[finishes #163369332] Modifed test tructure for ease understanding
